refactor dataset components
