+++ src/dsl/index.ts
@@ -1,5 +1,7 @@
 import { DSL_DashboardPage as Dashboard } from "../components/DashboardPage/Page";
-import {DSL_Dataset as Dataset, DSL_Transform as Transform, DSL_DataPreview as DataPreview} from "../components/Dataset/Dataset"
+import { DSL_Dataset as Dataset } from "../components/Dataset/Dataset";
+import { DSL_DataPreview as DataPreview } from "../components/Dataset/DataPreview";
+import { DSL_Transform as Transform} from "../components/Dataset/Transform";
 
 export {
     Dashboard,
